+++ packages/dashboard/src/pages/cluster-commands/index.ts
@@ -28,6 +28,7 @@ import "./clusters/basic-information-commands.js";
 import "./clusters/binding-commands.js";
 import "./clusters/chime-commands.js";
 import "./clusters/closure-control-commands.js";
+import "./clusters/closure-dimension-commands.js";
 import "./clusters/commodity-tariff-commands.js";
 import "./clusters/door-lock-commands.js";
 import "./clusters/icd-management-commands.js";